Convert sample app to Compose Multiplatform
