+++ sample/androidApp/.gitignore
