+++ sample/androidApp/proguard-rules.pro
