+++ sample/androidApp/src/main/AndroidManifest.xml
