+++ sample/androidApp/src/main/kotlin/io/github/elyesmansour/floatingTabBarSample/MainActivity.kt
@@ -1,0 +1,21 @@
+package io.github.elyesmansour.floatingTabBarSample
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
+import androidx.core.view.WindowCompat
+
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        enableEdgeToEdge()
+        WindowCompat.getInsetsController(window, window.decorView).isAppearanceLightStatusBars = false
+        WindowCompat.getInsetsController(window, window.decorView).isAppearanceLightNavigationBars = false
+
+        setContent {
+            App()
+        }
+    }
+}
