+++ sample/androidApp/src/main/res/values/colors.xml
