+++ sample/androidApp/src/main/res/values/strings.xml
