+++ sample/iosApp/Configuration/Config.xcconfig
@@ -1,0 +1,3 @@
+TEAM_ID=
+BUNDLE_ID=io.github.elyesmansour.floatingTabBarSample
+APP_NAME=FloatingTabBarSample
