+++ sample/iosApp/iosApp/ContentView.swift
@@ -1,0 +1,18 @@
+import UIKit
+import SwiftUI
+import SampleKit
+
+struct ComposeView: UIViewControllerRepresentable {
+    func makeUIViewController(context: Context) -> UIViewController {
+        MainViewControllerKt.MainViewController()
+    }
+
+    func updateUIViewController(_ uiViewController: UIViewController, context: Context) {}
+}
+
+struct ContentView: View {
+    var body: some View {
+        ComposeView()
+            .ignoresSafeArea(.all)
+    }
+}
