+++ sample/iosApp/iosApp/iOSApp.swift
@@ -1,0 +1,11 @@
+import SwiftUI
+
+@main
+struct iOSApp: App {
+    var body: some Scene {
+        WindowGroup {
+            ContentView()
+                .ignoresSafeArea(.all)
+        }
+    }
+}
