+++ sample/shared/src/commonMain/kotlin/io/github/elyesmansour/floatingTabBarSample/ui/theme/Color.kt
