+++ sample/shared/src/iosMain/kotlin/io/github/elyesmansour/floatingTabBarSample/MainViewController.kt
@@ -1,0 +1,7 @@
+package io.github.elyesmansour.floatingTabBarSample
+
+import androidx.compose.ui.window.ComposeUIViewController
+import platform.UIKit.UIViewController
+
+@Suppress("unused", "FunctionName")
+fun MainViewController(): UIViewController = ComposeUIViewController { App() }
