--- sample/src/main/java/io/github/elyesmansour/floatingTabBarSample/MainActivity.kt
@@ -1,24 +1,0 @@
-package io.github.elyesmansour.floatingTabBarSample
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.activity.enableEdgeToEdge
-import androidx.core.view.WindowCompat
-import io.github.elyesmansour.floatingTabBarSample.ui.theme.FloatingTabBarTheme
-
-class MainActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        enableEdgeToEdge()
-        WindowCompat.getInsetsController(window, window.decorView).isAppearanceLightStatusBars = false
-        WindowCompat.getInsetsController(window, window.decorView).isAppearanceLightNavigationBars = false
-
-        setContent {
-            FloatingTabBarTheme {
-                PlantSky()
-            }
-        }
-    }
-}
